[INFRA, MISC, TEST] move `detect_variants_in_alignment_file()` & update / create tests & organize includes
