+++ include/variant_detection/variant_detection.hpp
@@ -3,7 +3,7 @@
 #include <seqan3/std/filesystem>    // for filesystem
 #include <vector>
 
-#include "method_enums.hpp"         // for enum clustering_methods and refinement_methods
+#include "method_enums.hpp"         // for enum detection_methods, clustering_methods and refinement_methods
 #include "structures/junction.hpp"  // for class Junction
 
 //TODO (irallia): move this back to the iGenVar.cpp with the PR #104
@@ -47,31 +47,3 @@ void detect_junctions_in_long_reads_sam_file(std::vector<Junction> & junctions,
                                              clustering_methods const & clustering_method,
                                              refinement_methods const & refinement_method,
                                              uint64_t const min_var_length);
-
-/*! \brief Detects genomic variants by analyzing an alignment file (sam/bam). The detected
- *         variants are printed to a given file or stdout and insertion alleles are stored in a fasta file.
- *
- * \param[in]   args - command line arguments:\n
- *                     **args.alignment_short_reads_file_path** - short reads input file, path to the sam/bam file\n
- *                     **args.alignment_long_reads_file_path** - long reads input file, path to the sam/bam file\n
- *                     **args.output_file_path** output file - path for the VCF file
- *                     **args.methods** - list of methods for detecting junctions (1: cigar_string,
- *                                                                             2: split_read,
- *                                                                             3: read_pairs,
- *                                                                             4: read_depth)\n
- *                     **args.clustering_method** method for clustering junctions (0: simple_clustering
- *                                                                             1: hierarchical_clustering,
- *                                                                             2: self-balancing_binary_tree,
- *                                                                             3: candidate_selection_based_on_voting)\n
- *                     **args.refinement_method** method for refining junctions (0: no_refinement,
- *                                                                           1: sViper_refinement_method,
- *                                                                           2: sVirl_refinement_method)\n
- *                     **args.min_var_length** - minimum length of variants to detect (default 30 bp)\n
- *
- * \details Detects novel junctions from read alignment records using different detection methods.
- *          The junctions are clustered using one of several clustering methods.
- *          Then, the junction clusters are refined using one of several refinement methods.
- *          Finally, the refined junction clusters are categorized into different variant classes
- *          and output in VCF format.
- */
-void detect_variants_in_alignment_file(cmd_arguments const & args);
