+++ src/variant_detection/variant_output.cpp
@@ -4,7 +4,7 @@
 #include "variant_detection/variant_output.hpp"
 #include "variant_parser/variant_record.hpp"    // for class variant_header
 
-void find_and_output_variant(std::vector<Cluster> const & clusters, std::ostream & out_stream)
+void find_and_output_variants(std::vector<Cluster> const & clusters, std::ostream & out_stream)
 {
     variant_header header{};
     header.set_fileformat("VCFv4.3");
@@ -60,12 +60,12 @@ void find_and_output_variant(std::vector<Cluster> const & clusters, std::ostream
 }
 
 //!\overload
-void find_and_output_variant(std::vector<Cluster> const & clusters,
+void find_and_output_variants(std::vector<Cluster> const & clusters,
                              std::filesystem::path const & output_file_path)
 {
     if (output_file_path.empty())
     {
-        find_and_output_variant(clusters, std::cout);
+        find_and_output_variants(clusters, std::cout);
     }
     else
     {
@@ -74,7 +74,7 @@ void find_and_output_variant(std::vector<Cluster> const & clusters,
         {
             throw std::runtime_error{"Could not open file '" + output_file_path.string() + "' for reading."};
         }
-        find_and_output_variant(clusters, out_file);
+        find_and_output_variants(clusters, out_file);
         out_file.close();
     }
 }
