+++ test/api/refinement_test.cpp
@@ -1,0 +1,19 @@
+#include <gtest/gtest.h>
+
+// #include "modules/refinment/sViper_refinement_method.hpp"  // for the sViper refinement method
+
+/* -------- refinement methods tests -------- */
+
+// TODO (irallia): these tests should be implemented when the associated refinement methods are implemented
+
+// TEST(junction_detection, refinement_method_sViper)
+// {
+//     testing::internal::CaptureStdout();
+
+//     std::vector<Junction> junctions{};
+//     std::vector<Cluster> resulting_clusters{};
+//     sViper_refinement_method(junctions, clusters);
+
+//     std::vector<Cluster> expected_clusters{};
+//     EXPECT_EQ(expected_clusters, resulting_clusters);
+// }
