+++ test/cli/cli_test.hpp
@@ -1,3 +1,5 @@
+#pragma once
+
 #include <gtest/gtest.h>
 
 #include <cstdlib>               // system calls
@@ -7,8 +9,6 @@
 
 // Include the EXPECT_RANGE_EQ macro for better information if range elements differ.
 #include <seqan3/test/expect_range_eq.hpp>
-
-#pragma once
 
 // Provides functions for CLI test implementation.
 struct cli_test : public ::testing::Test
